requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.8
Deprecated==1.2.13
Django==3.2.6
django-cors-headers==3.8.0
django-environ==0.4.5
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
h11==0.12.0
idna==3.2
jwcrypto==1.0
oauthlib==3.1.1
psycopg2-binary==2.9.1
pycparser==2.20
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
uvicorn==0.14.0
whitenoise==5.3.0
wrapt==1.12.1
